requirements.txt
# Library dependencies for the python code.  You need to install these with
# First run: conda install pytorch=0.4.1 torchvision=0.2.1 -c pytorch
# Next `pip install -r requirements.txt` before you can run this.
#

Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
tensorboardX
pillow
numpy
opencv-contrib-python
networkx
